Continue dragging the orgchart, even if the mouse leaves the limits of the chart
If a user quickly drags the orgchart, so that the mouse leaves the boundaries
of the chart, the chart will stop dragging where the cursor left it.  This is
caused by the touch/mouse move events being bound to the chart, rather than the document.
By binding to the document, even if the cursor leaves the boundaries of the chart, the
chart will still be dragged to the cursors current location.