Drag-and-drop nodes on touch devices
Add support for drag-and-drop of nodes on touch devices.  Not 100% complete, as it does not include graphical effects during the 'drag' phase.  Tested on Chrome dev tools with touch simulator on Win 10, Safari on Apple iPad, and Chrome on SurfacePro.

This was originally developed against OrgChart version 2.0.8, so I've done a quick update based on today's master branch.  Apologies if the indentation has gone a bit wrong (Visual Studio is set up for my typical style, and I think it can't cope with OrgChart.js being different!).