Ensures that indeed options.compact() prevails over data.compact
This again is relaed to using oc.init(...) multiple times (which is normal when you are editing online a chart).

When the chart is created, each node get potentially the data.compact = true value. This is NOT reset currently if you define options.compact to be false and then perform a reinit. The current implementation only can "set" the value to true.

This change, makes (in line with the documentation), sets the values of the nodes in line with options.compact.